Replace changesets with pnpm native release management
