+++ package.json
@@ -18,7 +18,7 @@
     "fix:prettier": "prettier --write \"**/*.{md,yaml,yml}\"",
     "generate": "pnpm run /^generate:/",
     "generate:message-names": "node scripts/generate-message-names.ts",
-    "test": "node --test \"src/**/*.test.ts\" \"scripts/**/*.test.ts\""
+    "test": "node --test"
   },
   "dependencies": {
     "@base-ui/react": "1.6.0",
@@ -53,8 +53,6 @@
   },
   "devDependencies": {
     "@biomejs/biome": "2.5.3",
-    "@changesets/changelog-github": "0.7.0",
-    "@changesets/cli": "2.31.0",
     "@playwright/test": "1.61.1",
     "@types/css-tree": "2.3.11",
     "@types/fs-extra": "11.0.4",
@@ -73,7 +71,7 @@
     "prettier": "3.9.5",
     "typescript": "7.0.2"
   },
-  "packageManager": "pnpm@11.11.0",
+  "packageManager": "pnpm@11.15.1",
   "devEngines": {
     "runtime": {
       "name": "node",
